Packages/LavaSerpent/pyproject.toml
[project]
name = "lava-serpent"
version = "0.1.2"
description = ""
authors = [{ name = "alexstrnik", email = "alex.str.nik@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow-avif-plugin (>=1.5.2,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "rectangle-packer (>=2.0.4,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "lava_serpent" }]

[tool.poetry.scripts]
lava-serpent = "lava_serpent.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
